% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwl-package.R
\docType{package}
\name{cwl}
\alias{cwl}
\alias{cwl-package}
\title{cwl - Common workflow language support}
\description{
Parse and work with the common workflow language (cwl). See
\url{http://www.commonwl.org}.
}

