% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello boilerplate function}
\usage{
hello()
}
\value{
NULL. Called only for the side effect of messaging the console.
}
\description{
Description boilerplate for the hello function.
}
\examples{
hello()
#> [1] "Hello, world!"

}

