% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{readCwlData}
\alias{readCwlData}
\title{Read in a cwl tool data file}
\usage{
readCwlData(file)
}
\arguments{
\item{file}{The name of the YAML data file to read in}
}
\value{
The data as an object that can be used by a runner. This object is
essentially just a named list with each key being an element of the yaml
file.
}
\description{
Reads in an cwl data file (in yaml format) and returns an S3 "cwlDataYml"
object. This object is essentially just the yaml file read in as a named
list, and will work as a list if used as such. Parsing the yaml file is
done via the CRAN "yaml" package yaml.load_file function.
}
\examples{
file <- tempfile()
data <- c( "message: Hello world!", "" )
writeLines(data, file)

cwlDataObj <- readCwlData(file)
class(cwlDataObj)
#> [1] "cwlDataYml" "list"

length(cwlDataObj)
#> [1] 1

cwlDataObj[1]
#> $message
#> [1] "Hello world!"

}

