% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{readCwlTool}
\alias{readCwlTool}
\title{Read in a tool cwl file}
\usage{
readCwlTool(file)
}
\arguments{
\item{file}{The name of the cwl tool file to read in}
}
\value{
The tool.cwl as an object that can be used by a runner. This object is
essentially just a named list with each key being an element of the cwl yaml
file.
}
\description{
Reads in a tool cwl file (in yaml format) and returns an S3 "cwlTool"
object. This object is essentially just the cwl yaml file read in as a named
list, and will work as a list if used as such. Parsing the cwl yaml file is
done via the CRAN "yaml" package yaml.load_file function.
}
\examples{
# TODO: Hello world example
}

