% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{readCwlData}
\alias{readCwlData}
\title{Read in a cwl tool data file}
\usage{
readCwlData(file)
}
\arguments{
\item{file}{The name of the YAML data file to read in}
}
\value{
The data as an object that can be used by a runner.
}
\description{
Read in a cwl tool data file
}
\examples{
helloWorldDataFileName <- tempfile()
helloWorldData <- c( "message: Hello world!", "" )

writeLines(helloWorld, helloWorldDataFileName)

}

